+++ tests/functional/disaster-recovery/regional-dr/conftest.py
@@ -4,7 +4,6 @@ import os
 from ocs_ci.utility import templating
 import pytest
 import time
-import yaml
 
 from ocs_ci.framework import config
 from ocs_ci.helpers.virtctl import get_virtctl_tool
@@ -15,8 +14,10 @@ from ocs_ci.ocs.utils import get_non_acm_cluster_config
 from ocs_ci.utility.utils import (
     exec_cmd,
     run_cmd,
-    clone_repo,
-    wait_for_machineconfigpool_status,
+)
+from ocs_ci.helpers.dr_helpers import (
+    _apply_itms_to_managed_clusters,
+    _generate_rdr_mirror_images,
 )
 from ocs_ci.ocs.exceptions import (
     CommandFailed, ResourceNotFoundError,
@@ -51,7 +52,7 @@ def pytest_collection_modifyitems(items):
                 items.remove(item)
 
 
-@pytest.fixture(autouse=True, scope="session")
+@pytest.fixture(scope="session", autouse=True)
 def check_subctl_cli():
     if config.MULTICLUSTER.get("multicluster_mode") != constants.RDR_MODE:
         return
@@ -307,162 +308,3 @@ def mirror_rdr_images():
         error_msg = f"ITMS file not found at expected location: {itms_file_path}"
         log.error(error_msg)
         raise ResourceNotFoundError(error_msg)
-
-
-def _generate_rdr_mirror_images():
-    """
-    Extract and return list of container images from RDR workload repository.
-
-    Returns:
-        list: List of container image strings for mirroring in disconnected environments
-    """
-    # Check whether deployment is disconnected
-    if config.DEPLOYMENT.get("disconnected"):
-
-        workload_repo_url = config.ENV_DATA["dr_workload_repo_url"]
-        log.info(f"Repo used: {workload_repo_url}")
-        workload_repo_branch = config.ENV_DATA["dr_workload_repo_branch"]
-        clone_repo(
-            url=workload_repo_url,
-            location="/tmp/ocs-workload-repo",
-            branch=workload_repo_branch,
-        )
-
-        # List all Kubernetes container images under rdr/ folder
-        rdr_path = "/tmp/ocs-workload-repo/rdr"
-        if os.path.exists(rdr_path):
-            log.info(f"Extracting Kubernetes container images from {rdr_path}:")
-            images_found = set()
-
-            for root, dirs, files in os.walk(rdr_path):
-                for file in files:
-                    # Process YAML files
-                    if file.lower().endswith((".yaml", ".yml")):
-                        yaml_file_path = os.path.join(root, file)
-                        relative_path = os.path.relpath(yaml_file_path, rdr_path)
-
-                        try:
-                            with open(yaml_file_path, "r") as f:
-                                yaml_content = yaml.safe_load_all(f)
-
-                                for doc in yaml_content:
-                                    if doc:
-                                        # Extract images from the YAML document
-                                        images = extract_images_from_yaml(doc)
-                                        for image in images:
-                                            if image not in images_found:
-                                                images_found.add(image)
-                                                log.info(
-                                                    f"  - {image} (from {relative_path})"
-                                                )
-                        except Exception as e:
-                            log.warning(f"Failed to parse {relative_path}: {e}")
-
-            # Convert set to sorted list for consistent ordering
-            image_list = sorted(list(images_found))
-
-            if image_list:
-                log.info(f"Total unique container images found: {len(image_list)}")
-                log.info("Images ready for mirroring in disconnected environment")
-            else:
-                log.warning("No container images found in YAML files")
-            log.info(image_list)
-            return image_list
-        else:
-            log.warning(f"RDR path does not exist: {rdr_path}")
-            return []
-
-    return []
-
-
-def _apply_itms_to_managed_clusters(itms_file_path):
-    """
-    Apply itms configuration to all managed clusters and wait for MCP to complete.
-
-    Args:
-        itms_file_path (str): Path to the itms-oc-mirror.yaml file
-    """
-    log.info("Applying itms to managed clusters")
-
-    # Get all managed cluster configs
-    managed_clusters = get_non_acm_cluster_config()
-
-    if not managed_clusters:
-        log.warning("No managed clusters found")
-        return
-
-    # Store original context to restore later
-    original_ctx = config.cur_index
-
-    try:
-        for cluster_config in managed_clusters:
-            cluster_name = cluster_config.ENV_DATA.get("cluster_name", "unknown")
-            log.info(f"Applying itms to managed cluster: {cluster_name}")
-
-            # Switch to the managed cluster context
-            cluster_index = cluster_config.MULTICLUSTER.get("multicluster_index")
-            if cluster_index is not None:
-                config.switch_ctx(cluster_index)
-
-                try:
-                    # Apply the itms file
-                    run_cmd(f"oc apply -f {itms_file_path}")
-                    log.info(f"Successfully applied itms to {cluster_name}")
-
-                    # Wait for MachineConfigPool to complete
-                    log.info(
-                        f"Waiting for MachineConfigPool to complete on {cluster_name}"
-                    )
-                    wait_for_machineconfigpool_status("all", timeout=1800)
-                    log.info(f"MachineConfigPool update completed on {cluster_name}")
-
-                except CommandFailed as e:
-                    log.error(f"Failed to apply itms to {cluster_name}: {e}")
-                    raise
-                except Exception as e:
-                    log.error(f"Error during itms application to {cluster_name}: {e}")
-                    raise
-            else:
-                log.warning(f"Could not find cluster index for {cluster_name}")
-
-    finally:
-        # Restore original context
-        config.switch_ctx(original_ctx)
-        log.info("Restored original cluster context")
-
-
-def extract_images_from_yaml(obj, images=None):
-    """
-    Recursively extract container image references from a YAML object.
-    Extracts values from 'image', 'url', and 'value' keys that contain image references.
-
-    Args:
-        obj: YAML object (dict, list, or primitive)
-        images: Set to collect images (created if None)
-
-    Returns:
-        set: Set of container image strings
-    """
-    if images is None:
-        images = set()
-
-    if isinstance(obj, dict):
-        # Check for 'image', 'url', and 'value' keys that contain image references
-        for key in ["image", "url", "value"]:
-            if key in obj and isinstance(obj[key], str):
-                # Only add if it looks like a container image reference
-                # (contains registry path or common image patterns)
-                value = obj[key].strip()
-                if value and ("/" in value or ":" in value):
-                    images.add(value)
-
-        # Recursively process all values
-        for value in obj.values():
-            extract_images_from_yaml(value, images)
-
-    elif isinstance(obj, list):
-        # Recursively process all items
-        for item in obj:
-            extract_images_from_yaml(item, images)
-
-    return images
